% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MouseDLN.SCE}
\alias{MouseDLN.SCE}
\title{\linkS4class{SingleCellExperiment} object containing mouse lymph node
single-cell RNA-seq data from Lopez, Li, and Keren-Shaul et al., 2022}
\format{
An object of class \code{SingleCellExperiment} with 12854 rows and 14989 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE173778}
}
\usage{
MouseDLN.SCE
}
\description{
\linkS4class{SingleCellExperiment} object containing mouse lymph node
single-cell RNA-seq data generated from Lopez, Li, and Keren-Shaul et al.,
2022. Raw data are available here:
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE173778}.
}
\references{
Lopez, R., Li, B., Keren-Shaul, H. et al. DestVI identifies
continuums of cell types in spatial transcriptomics data. Nat Biotechnol
40, 1360–1369 (2022). doi: 10.1038/s41587-022-01272-8
}
\keyword{datasets}
