% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MouseDLN.ST}
\alias{MouseDLN.ST}
\title{\linkS4class{SpatialExperiment} object containing mouse lymph node spatial
transcriptomics data (10X Visium)  from Lopez, Li, and Keren-Shaul et al.,
2022}
\format{
An object of class \code{SpatialExperiment} with 13948 rows and 1092 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE173778}
}
\usage{
MouseDLN.ST
}
\description{
\linkS4class{SpatialExperiment} object containing mouse lymph node spatial
transcriptomics data (10X Visium)  from Lopez, Li, and Keren-Shaul et al.,
2022. Raw data are available here:
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE173778}.
}
\references{
Lopez, R., Li, B., Keren-Shaul, H. et al. DestVI identifies
continuums of cell types in spatial transcriptomics data. Nat Biotechnol
40, 1360–1369 (2022). doi: 10.1038/s41587-022-01272-8
}
\keyword{datasets}
